+++ lib/utils/utils.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 
 import 'package:app_fleet/app/config/domain/workspace_entity.dart';
 import 'package:app_fleet/core/app_bug_report.dart';
+import 'package:app_fleet/core/logger.dart';
 import 'package:app_fleet/main.dart';
 import 'package:flutter/material.dart';
 
@@ -40,15 +41,18 @@ Map<String, Color> accentColorMap = {
 void executeWorkspaceSwitcher(int workspaceNumber) {
   try {
     Process.runSync(
-      combinePath([
+      combineHomePath([
         '.config',
+        "app-fleet",
         'workspace-switcher.sh',
       ]),
       [workspaceNumber.toString()],
     );
   } catch (error, stackTrace) {
-    debugPrintApp(
-        "[WorkspaceSwitcher] Got an error when trying to switch to workspace number $workspaceNumber");
+    prettyLog(
+        tag: "WorkspaceSwitcher",
+        value:
+            "Got an error when trying to switch to workspace number $workspaceNumber");
     AppBugReport.createReport(
       message:
           "Got an error when trying to switch to workspace number $workspaceNumber",
@@ -63,15 +67,18 @@ void executeWorkspaceSwitcher(int workspaceNumber) {
 Future<void> executeProcessExecutor(String command) async {
   try {
     await Process.start(
-      combinePath([
+      combineHomePath([
         ".config",
+        "app-fleet",
         'unix-process-executor.sh',
       ]),
       [command],
     );
   } catch (error, stackTrace) {
-    debugPrintApp(
-        "[ProcessExecutor] Got an error when trying to run command: \"$command\"");
+    prettyLog(
+        tag: "ProcessExecutor",
+        value: "Got an error when trying to run command: \"$command\"",
+        type: DebugType.error);
     AppBugReport.createReport(
       message: "Got an error when trying to run command: \"$command\"",
       source: "`utils.dart` - `executeProcessExecutor()`",
@@ -86,8 +93,9 @@ Future<void> executeProcessExecutor(String command) async {
 String getSystemTheme() {
   try {
     final result = Process.runSync(
-      combinePath([
+      combineHomePath([
         '.config',
+        "app-fleet",
         'theme-detector.sh',
       ]),
       [],
@@ -95,8 +103,11 @@ String getSystemTheme() {
     int exitCode = result.exitCode;
     return exitCode == 1 ? 'dark' : 'light';
   } on Exception {
-    debugPrintApp(
-        "[WorkspaceSwitcher] Got an error when trying to identify System Theme");
+    prettyLog(
+      tag: "WorkspaceSwitcher",
+      value: "Got an error when trying to identify System Theme",
+      type: DebugType.error,
+    );
   }
   // falling back to light if any error occurs
   return 'light';
@@ -107,10 +118,16 @@ String combinePath(List<String> locations, {bool absolute = false}) {
   return absolute ? File(path).absolute.path : path;
 }
 
+String combineHomePath(List<String> locations, {bool absolute = false}) {
+  locations.insert(0, Platform.environment['HOME']!);
+  return combinePath(locations, absolute: absolute);
+}
+
 void mkdir(String path, String logMessage) {
   var dir = Directory(path);
   if (!dir.existsSync()) {
     dir.createSync();
+    prettyLog(value: logMessage);
   }
 }
 
@@ -134,16 +151,16 @@ void switchWorkspace(int number) {
 }
 
 String getBugReportPath(String reportID) {
-  return "file://${Platform.environment['HOME']}/app-fleet/.config/bug-reports/$reportID.md";
+  return "file://${Platform.environment['HOME']}/.config/app-fleet/bug-reports/$reportID.md";
 }
 
 String getWorkspacePath(String workspaceName) {
-  return "file://${Platform.environment['HOME']}/app-fleet/.config/workspaces/$workspaceName.json";
+  return "file://${Platform.environment['HOME']}/.config/app-fleet/workspaces/$workspaceName.json";
 }
 
 bool launcherModeCapable() {
   return debugMode ||
-      File('${Platform.environment['HOME']}/app-fleet/.config/app-settings.json')
+      File('${Platform.environment['HOME']}/.config/app-fleet/app-settings.json')
           .existsSync();
 }